Refactor transition words assessment
