+++ packages/yoastseo/spec/scoring/assessments/readability/transitionWordsAssessmentSpec.js
@@ -1,6 +1,6 @@
 import DefaultResearcher from "../../../../src/languageProcessing/languages/_default/Researcher";
 import EnglishResearcher from "../../../../src/languageProcessing/languages/en/Researcher";
-import transitionWordsAssessment from "../../../../src/scoring/assessments/readability/transitionWordsAssessment.js";
+import TransitionWordsAssessment from "../../../../src/scoring/assessments/readability/TransitionWordsAssessment.js";
 import Paper from "../../../../src/values/Paper.js";
 import Factory from "../../../specHelpers/factory.js";
 import Mark from "../../../../src/values/Mark.js";
@@ -9,7 +9,7 @@ const i18n = Factory.buildJed();
 describe( "An assessment for transition word percentage", function() {
 	it( "returns the score for 0% of the sentences with transition words", function() {
 		const mockPaper = new Paper();
-		const assessment = transitionWordsAssessment.getResult( mockPaper, Factory.buildMockResearcher( { totalSentences: 10,
+		const assessment = new TransitionWordsAssessment().getResult( mockPaper, Factory.buildMockResearcher( { totalSentences: 10,
 			transitionWordSentences: 0 } ), i18n );
 
 		expect( assessment.getScore() ).toEqual( 3 );
@@ -20,7 +20,7 @@ describe( "An assessment for transition word percentage", function() {
 
 	it( "returns the score for a paper with text but no sentences (e.g. only images)", function() {
 		const mockPaper = new Paper();
-		const assessment = transitionWordsAssessment.getResult( mockPaper, Factory.buildMockResearcher( { totalSentences: 0,
+		const assessment = new TransitionWordsAssessment().getResult( mockPaper, Factory.buildMockResearcher( { totalSentences: 0,
 			transitionWordSentences: 0 } ), i18n );
 
 		expect( assessment.getScore() ).toEqual( 3 );
@@ -31,7 +31,7 @@ describe( "An assessment for transition word percentage", function() {
 
 	it( "returns the score for 10.0% of the sentences with transition words", function() {
 		const mockPaper = new Paper();
-		const assessment = transitionWordsAssessment.getResult( mockPaper, Factory.buildMockResearcher( { totalSentences: 10,
+		const assessment = new TransitionWordsAssessment().getResult( mockPaper, Factory.buildMockResearcher( { totalSentences: 10,
 			transitionWordSentences: 1 } ), i18n );
 
 		expect( assessment.getScore() ).toEqual( 3 );
@@ -42,7 +42,7 @@ describe( "An assessment for transition word percentage", function() {
 	} );
 	it( "returns the score for 20.0% of the sentences with transition words", function() {
 		const mockPaper = new Paper();
-		const assessment = transitionWordsAssessment.getResult( mockPaper, Factory.buildMockResearcher( { totalSentences: 5,
+		const assessment = new TransitionWordsAssessment().getResult( mockPaper, Factory.buildMockResearcher( { totalSentences: 5,
 			transitionWordSentences: 1 } ), i18n );
 
 		expect( assessment.getScore() ).toEqual( 6 );
@@ -53,7 +53,7 @@ describe( "An assessment for transition word percentage", function() {
 	} );
 	it( "returns the score for 25.0% of the sentences with transition words", function() {
 		const mockPaper = new Paper();
-		const assessment = transitionWordsAssessment.getResult( mockPaper, Factory.buildMockResearcher( { totalSentences: 4,
+		const assessment = new TransitionWordsAssessment().getResult( mockPaper, Factory.buildMockResearcher( { totalSentences: 4,
 			transitionWordSentences: 1 } ), i18n );
 
 		expect( assessment.getScore() ).toEqual( 6 );
@@ -64,7 +64,7 @@ describe( "An assessment for transition word percentage", function() {
 	} );
 	it( "returns the score for 35.0% of the sentences with transition words", function() {
 		const mockPaper = new Paper();
-		const assessment = transitionWordsAssessment.getResult( mockPaper, Factory.buildMockResearcher( { totalSentences: 20,
+		const assessment = new TransitionWordsAssessment().getResult( mockPaper, Factory.buildMockResearcher( { totalSentences: 20,
 			transitionWordSentences: 7 } ), i18n );
 		expect( assessment.getScore() ).toEqual( 9 );
 		expect( assessment.getText() ).toEqual( "<a href='https://yoa.st/34z' target='_blank'>Transition words</a>: Well done!" );
@@ -72,7 +72,7 @@ describe( "An assessment for transition word percentage", function() {
 	} );
 	it( "returns the score for 40% sentences with transition words", function() {
 		const mockPaper = new Paper();
-		const assessment = transitionWordsAssessment.getResult( mockPaper, Factory.buildMockResearcher( { totalSentences: 10,
+		const assessment = new TransitionWordsAssessment().getResult( mockPaper, Factory.buildMockResearcher( { totalSentences: 10,
 			transitionWordSentences: 4 } ), i18n );
 		expect( assessment.getScore() ).toEqual( 9 );
 		expect( assessment.getText() ).toEqual( "<a href='https://yoa.st/34z' target='_blank'>Transition words</a>: Well done!" );
@@ -81,7 +81,7 @@ describe( "An assessment for transition word percentage", function() {
 
 	it( "returns the score for 47% sentences with transition words", function() {
 		const mockPaper = new Paper();
-		const assessment = transitionWordsAssessment.getResult( mockPaper, Factory.buildMockResearcher( { totalSentences: 100,
+		const assessment = new TransitionWordsAssessment().getResult( mockPaper, Factory.buildMockResearcher( { totalSentences: 100,
 			transitionWordSentences: 47 } ), i18n );
 		expect( assessment.getScore() ).toEqual( 9 );
 		expect( assessment.getText() ).toEqual( "<a href='https://yoa.st/34z' target='_blank'>Transition words</a>: Well done!" );
@@ -90,7 +90,7 @@ describe( "An assessment for transition word percentage", function() {
 
 	it( "returns the score for 66.7% of the sentences with transition words", function() {
 		const mockPaper = new Paper();
-		const assessment = transitionWordsAssessment.getResult( mockPaper, Factory.buildMockResearcher( { totalSentences: 3,
+		const assessment = new TransitionWordsAssessment().getResult( mockPaper, Factory.buildMockResearcher( { totalSentences: 3,
 			transitionWordSentences: 2 } ), i18n );
 
 		expect( assessment.getScore() ).toEqual( 9 );
@@ -100,7 +100,7 @@ describe( "An assessment for transition word percentage", function() {
 
 	it( "is not applicable for empty papers", function() {
 		const mockPaper = new Paper();
-		const assessment = transitionWordsAssessment.isApplicable( mockPaper, new EnglishResearcher( mockPaper ) );
+		const assessment = new TransitionWordsAssessment().isApplicable( mockPaper, new EnglishResearcher( mockPaper ) );
 		expect( assessment ).toBe( false );
 	} );
 
@@ -109,7 +109,7 @@ describe( "An assessment for transition word percentage", function() {
 			"Ut vim eius porro labore. Id quem civibus sit. Sed no primis urbanitas, aperiri laboramus voluptatibus ei per. Esse consul possim " +
 			"duo eu, eu duo natum ferri libris. Tritani percipit interpretaris ne ius. Mel prima definitionem eu, partem labores vim at. " +
 			"Prompta vivendum usu te. Indoctum philosophia definitiones usu ad, cum quodsi alienum et. " );
-		const assessment = transitionWordsAssessment.isApplicable( mockPaper, new EnglishResearcher( mockPaper ) );
+		const assessment = new TransitionWordsAssessment().isApplicable( mockPaper, new EnglishResearcher( mockPaper ) );
 		expect( assessment ).toBe( false );
 	} );
 
@@ -128,13 +128,13 @@ describe( "An assessment for transition word percentage", function() {
 			"aliquando ne vix. Fabulas voluptua eu vel. Ceteros euripidis has cu. Pro ea esse ignota perfecto, ius noluisse liberavisse ei. " +
 			"Has possim mediocritatem in. Paulo alienum accusamus pro cu, magna labore sit ad. Sumo paulo sea in, cum te latine " +
 			"labores inciderint.", { locale: "en_US" } );
-		const assessment = transitionWordsAssessment.isApplicable( mockPaper, new EnglishResearcher( mockPaper ) );
+		const assessment = new TransitionWordsAssessment().isApplicable( mockPaper, new EnglishResearcher( mockPaper ) );
 		expect( assessment ).toBe( true );
 	} );
 
 	it( "is not applicable when used with a non-supported researcher, e.g. the default researcher", function() {
 		const mockPaper = new Paper( "This is a string", { locale: "xx_YY" } );
-		const assessment = transitionWordsAssessment.isApplicable( mockPaper, new DefaultResearcher( mockPaper ) );
+		const assessment = new TransitionWordsAssessment().isApplicable( mockPaper, new DefaultResearcher( mockPaper ) );
 		expect( assessment ).toBe( false );
 	} );
 } );
@@ -148,13 +148,13 @@ describe( "A test for marking sentences containing a transition word", function(
 			new Mark( { original: "This sentence is marked, because it contains a transition word.", marked: "<yoastmark " +
 					"class='yoast-text-mark'>This sentence is marked, because it contains a transition word.</yoastmark>" } ),
 		];
-		expect( transitionWordsAssessment.getMarks( paper, transitionWords ) ).toEqual( expected );
+		expect( new TransitionWordsAssessment().getMarks( paper, transitionWords ) ).toEqual( expected );
 	} );
 
 	it( "returns no markers if no sentences contain a transition word", function() {
 		const paper = new Paper( "This sentence is not marked." );
 		const transitionWords = Factory.buildMockResearcher( { sentenceResults: [ ] } );
 		const expected = [];
-		expect( transitionWordsAssessment.getMarks( paper, transitionWords ) ).toEqual( expected );
+		expect( new TransitionWordsAssessment().getMarks( paper, transitionWords ) ).toEqual( expected );
 	} );
 } );
